changed to use iihoc, removed unused code, removed object spread operator to keep eslint happy
