--- .babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "targets": {
-        "browsers": ["last 2 versions", "safari >= 7"]
-      }
-    }],
-    "react"
-  ]
-}
