+++ package.json
@@ -40,10 +40,6 @@
     "coverage": "rc-tools run coverage"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-react": "^6.24.1",
-    "eslint": "^3.19.0",
     "expect.js": "~0.3.1",
     "pre-commit": "1.x",
     "rc-tools": "5.x",
